+++ src/Repository/Loader/FileRepositoryLoader.php
@@ -1,5 +1,6 @@
 <?php
-/**
+
+/*
  * UserFrosting Support (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/support
@@ -25,6 +26,7 @@ abstract class FileRepositoryLoader
 
     /**
      * Create the loader.
+     *
      * @param string|string[] $paths
      */
     public function __construct($paths)
@@ -35,13 +37,16 @@ abstract class FileRepositoryLoader
     /**
      * Fetch content from a single file path.
      *
-     * @param  string $path
+     * @param string $path
+     *
      * @return array
      */
     abstract protected function parseFile($path);
 
     /**
      * Fetch and recursively merge in content from all file paths.
+     *
+     * @param bool $skipMissing
      *
      * @return array
      */
@@ -60,9 +65,11 @@ abstract class FileRepositoryLoader
     /**
      * Fetch content from a single file path.
      *
-     * @param  string                $path
-     * @param  bool                  $skipMissing True to ignore bad file paths.  If set to false, will throw an exception instead.
+     * @param string $path
+     * @param bool   $skipMissing True to ignore bad file paths.  If set to false, will throw an exception instead.
+     *
      * @throws FileNotFoundException
+     *
      * @return array
      */
     public function loadFile($path, $skipMissing = true)
@@ -76,11 +83,23 @@ abstract class FileRepositoryLoader
         }
 
         // If the file exists but is not readable, we always throw an exception.
-        if (!is_readable($path)) {
+        if (!$this->isReadable($path)) {
             throw new FileNotFoundException("The repository file '$path' exists, but it could not be read.");
         }
 
         return $this->parseFile($path);
+    }
+
+    /**
+     * Return if path is readable.
+     *
+     * @param string $path
+     *
+     * @return bool
+     */
+    protected function isReadable($path)
+    {
+        return is_readable($path);
     }
 
     /**
@@ -102,7 +121,7 @@ abstract class FileRepositoryLoader
      */
     public function prependPath($path)
     {
-        array_unshift($this->paths[], rtrim($path, '/\\'));
+        array_unshift($this->paths, rtrim($path, '/\\'));
 
         return $this;
     }