+++ src/Util/Util.php
@@ -1,5 +1,6 @@
 <?php
-/**
+
+/*
  * UserFrosting Support (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/support
@@ -10,7 +11,7 @@
 namespace UserFrosting\Support\Util;
 
 /**
- * Util Class
+ * Util Class.
  *
  * Static utility functions for UserFrosting.
  *
@@ -26,10 +27,13 @@ class Util
      * By default (if $throwException parameter is not set to true) returns false on failure.
      *
      * @see https://github.com/rockettheme/toolbox/blob/develop/ResourceLocator/src/UniformResourceLocator.php
-     * @param  string                  $uri
-     * @param  bool                    $throwException
-     * @param  bool                    $splitStream
+     *
+     * @param string $uri
+     * @param bool   $throwException
+     * @param bool   $splitStream
+     *
      * @throws \BadMethodCallException
+     *
      * @return string|array|bool
      */
     public static function normalizePath($uri, $throwException = false, $splitStream = false)
@@ -82,12 +86,19 @@ class Util
     /**
      * Removes a prefix from the beginning of a string, if a match is found.
      *
-     * @param  string $str    The string to process.
-     * @param  string $prefix The prefix to find and remove.
+     * @param string $str    The string to process.
+     * @param string $prefix The prefix to find and remove.
+     *
      * @return string
      */
     public static function stripPrefix($str, $prefix = '')
     {
+        // if string is same as prefix, return empty string
+        // Otherwise PHP 5.6 will return false
+        if ($str === $prefix) {
+            return '';
+        }
+
         if (substr($str, 0, strlen($prefix)) == $prefix) {
             $str = substr($str, strlen($prefix));
         }
@@ -98,12 +109,13 @@ class Util
     /**
      * Determine if a given string matches one or more regular expressions.
      *
-     * @param  string|array $patterns
-     * @param  string       $subject
-     * @param  array        &$matches
-     * @param  string       $delimiter
-     * @param  int          $flags
-     * @param  int          $offset
+     * @param string|array $patterns
+     * @param string       $subject
+     * @param array        &$matches
+     * @param string       $delimiter
+     * @param int          $flags
+     * @param int          $offset
+     *
      * @return bool
      */
     public static function stringMatches($patterns, $subject, array &$matches = null, $delimiter = '~', $flags = 0, $offset = 0)
@@ -112,7 +124,7 @@ class Util
         $result = false;
         foreach ((array) $patterns as $pattern) {
             $currMatches = [];
-            if ($pattern != '' && preg_match($delimiter . $pattern . $delimiter, $subject, $currMatches, $flags, $offset)) {
+            if ($pattern != '' && preg_match($delimiter.$pattern.$delimiter, $subject, $currMatches, $flags, $offset)) {
                 $result = true;
                 $matches[$pattern] = $currMatches;
             }
@@ -124,8 +136,9 @@ class Util
     /**
      * Recursively apply a callback to members of an array.
      *
-     * @param  array    $input
-     * @param  callback $callback
+     * @param array    $input
+     * @param callable $callback
+     *
      * @return array
      */
     public static function arrayFilterRecursive($input, $callback = null)