+++ tests/Repository/PathBuilderTest.php
@@ -1,9 +1,17 @@
 <?php
 
+/*
+ * UserFrosting Support (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/support
+ * @copyright Copyright (c) 2013-2019 Alexander Weissman
+ * @license   https://github.com/userfrosting/support/blob/master/LICENSE.md (MIT License)
+ */
+
 use PHPUnit\Framework\TestCase;
-use UserFrosting\UniformResourceLocator\ResourceLocator;
 use UserFrosting\Support\Repository\PathBuilder\SimpleGlobBuilder;
 use UserFrosting\Support\Repository\PathBuilder\StreamPathBuilder;
+use UserFrosting\UniformResourceLocator\ResourceLocator;
 
 class PathBuilderTest extends TestCase
 {
@@ -13,7 +21,7 @@ class PathBuilderTest extends TestCase
 
     public function setUp()
     {
-        $this->basePath = __DIR__ . '/data';
+        $this->basePath = __DIR__.'/data';
         $this->locator = new ResourceLocator($this->basePath);
 
         $this->locator->registerStream('owls');
@@ -34,13 +42,12 @@ class PathBuilderTest extends TestCase
 
         // Assert
         $this->assertEquals($paths, [
-            $this->basePath . '/core/owls/megascops.php',
-            $this->basePath . '/core/owls/tyto.php',
-            $this->basePath . '/account/owls/megascops.php',
-            $this->basePath . '/admin/owls/megascops.php'
+            $this->basePath.'/core/owls/megascops.php',
+            $this->basePath.'/core/owls/tyto.php',
+            $this->basePath.'/account/owls/megascops.php',
+            $this->basePath.'/admin/owls/megascops.php',
         ]);
     }
-
 
     public function testBuildPathsToFile()
     {
@@ -52,9 +59,9 @@ class PathBuilderTest extends TestCase
 
         // Assert
         $this->assertEquals([
-            $this->basePath . '/core/owls/megascops.php',
-            $this->basePath . '/account/owls/megascops.php',
-            $this->basePath . '/admin/owls/megascops.php'
+            $this->basePath.'/core/owls/megascops.php',
+            $this->basePath.'/account/owls/megascops.php',
+            $this->basePath.'/admin/owls/megascops.php',
         ], $paths);
     }
 }