+++ tests/Repository/data/account/owls/megascops.php
@@ -1,0 +1,15 @@
+<?php
+
+/*
+ * UserFrosting Support (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/support
+ * @copyright Copyright (c) 2013-2019 Alexander Weissman
+ * @license   https://github.com/userfrosting/support/blob/master/LICENSE.md (MIT License)
+ */
+
+return [
+    'voles' => [
+        'devoured' => 8,
+    ],
+];