+++ tests/Repository/data/core/owls/megascops.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * UserFrosting Support (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/support
+ * @copyright Copyright (c) 2013-2019 Alexander Weissman
+ * @license   https://github.com/userfrosting/support/blob/master/LICENSE.md (MIT License)
+ */
+
+return [
+    'voles' => [
+        'caught'   => 8,
+        'devoured' => 5,
+    ],
+    'plumage' => 'fluffy',
+    'chicks'  => 3,
+];