+++ tests/data/account/owls/megascops.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'voles' => [
-        'devoured' => 8
-    ]
-];