+++ tests/data/admin/owls/megascops.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'chicks' => 4
-];