+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.2.1]
+- Fix issue with FileRepositoryLoader's `prependPath` method.
+- Fix issue with `Util::stripPrefix` if prefix is same as string to process on PHP 5.6.
+- 100% test coverage
+
 ## [4.2.0]
 - Replaced `rockettheme/toolbox` with `userfrosting/uniformresourcelocator`.
 
@@ -24,6 +29,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## 4.0.0
 - Initial Release
 
+[4.2.1]: https://github.com/userfrosting/support/compare/4.2.0...4.2.1
 [4.2.0]: https://github.com/userfrosting/support/compare/4.1.3...4.2.0
 [4.1.3]: https://github.com/userfrosting/support/compare/4.1.2...4.1.3
 [4.1.2]: https://github.com/userfrosting/support/compare/4.1.1...4.1.2